fix: resolve MA0025/MA0047 warnings-as-errors surfaced by SDK 0.0.119
The 0.0.117->0.0.119 SDK bump (#563) corrected analyzer-config import/NoWarn
timing, newly activating the always-on baseline on demo/dashboard projects.

- Dashboard: throw NotSupportedException (not NotImplementedException) for the
  unsupported CircularBuffer ICollection ops and INodeDiscoveryProvider default
  interface methods (MA0025) + matching doc-comment cref updates
- Demo: move helper DbContexts out of the top-level-statement Program.cs files
  into their own namespaced files so they satisfy MA0047 (kept enforced)
- .editorconfig [tests/**]: suppress MA0025 (tests intentionally throw
  NotImplementedException to exercise exception-handler mapping)
